CHANGES TO THE WIDE SCREEN APPEARANCE
I have added flex to the h1, input and submit button, and also flex to "show tasks" buttons so they take as much space as h1 element, as well input and submit button line and they've been justified to space-around. Necessary items have been aligned to center (cross axis).
In the editing template (TodoItem.js) style of calc function has been applied for the input field (that shows up while editing), so the input field takes same amount of space as the previously mentioned elements. It didn't work when it was added in the CSS file.